Rename projectData module to project.
