+++ routes/feature.js
@@ -7,7 +7,7 @@ var express = require('express');
 var router = express.Router();
 
 var markdown = require('markdown').markdown;
-var getProjectData = require('../lib/specifications/projectData').getData;
+var getProjectData = require('../lib/specifications/project').getData;
 var getFileContents = require('../lib/specifications/repositoryTypes/git').getFileContents;
 var appConfig = require('../lib/configuration').get();
